test: Add `connectedCallback` invocation order test (#2764)
