+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightContainer/lightContainer.js
@@ -3,6 +3,10 @@ import { LightningElement } from 'lwc';
 export default class LightContainer extends LightningElement {
     static renderMode = 'light';
 
+    connectedCallback() {
+        window.timingBuffer.push(`lightContainer:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`lightContainer:disconnectedCallback`);
     }
