+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightParent/lightParent.js
@@ -3,6 +3,10 @@ import { LightningElement } from 'lwc';
 export default class LightParent extends LightningElement {
     static renderMode = 'light';
 
+    connectedCallback() {
+        window.timingBuffer.push(`lightParent:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`lightParent:disconnectedCallback`);
     }
