+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightShadowParent/lightShadowParent.js
@@ -3,6 +3,10 @@ import { LightningElement } from 'lwc';
 export default class LightShadowParent extends LightningElement {
     static renderMode = 'light';
 
+    connectedCallback() {
+        window.timingBuffer.push(`lightShadowContainer:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`lightShadowContainer:disconnectedCallback`);
     }
