+++ packages/integration-karma/test/rendering/callback-invocation-order/x/parent/parent.js
@@ -3,6 +3,10 @@ import { LightningElement, api } from 'lwc';
 export default class Parent extends LightningElement {
     @api name;
 
+    connectedCallback() {
+        window.timingBuffer.push(`parent:${this.name}:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`parent:${this.name}:disconnectedCallback`);
     }
