+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowContainer/shadowContainer.js
@@ -1,6 +1,10 @@
 import { LightningElement } from 'lwc';
 
 export default class ShadowContainer extends LightningElement {
+    connectedCallback() {
+        window.timingBuffer.push(`shadowContainer:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`shadowContainer:disconnectedCallback`);
     }
