+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowLightParent/shadowLightParent.js
@@ -1,6 +1,10 @@
 import { LightningElement } from 'lwc';
 
 export default class ShadowLightParent extends LightningElement {
+    connectedCallback() {
+        window.timingBuffer.push(`shadowLightParent:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`shadowLightParent:disconnectedCallback`);
     }
