+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowParent/shadowParent.js
@@ -1,6 +1,10 @@
 import { LightningElement } from 'lwc';
 
 export default class ShadowParent extends LightningElement {
+    connectedCallback() {
+        window.timingBuffer.push(`shadowParent:connectedCallback`);
+    }
+
     disconnectedCallback() {
         window.timingBuffer.push(`shadowParent:disconnectedCallback`);
     }
