+++ packages/integration-karma/test/rendering/callback-invocation-order/x/toggleParent/toggleParent.js
@@ -1,20 +1,11 @@
 import { LightningElement } from 'lwc';
 
 export default class ToggleParent extends LightningElement {
-    constructor() {
-        super();
-        window.timingBuffer.push('parent:constructor');
-    }
-
     connectedCallback() {
         window.timingBuffer.push('parent:connectedCallback');
     }
 
     disconnectedCallback() {
         window.timingBuffer.push('parent:disconnectedCallback');
-    }
-
-    renderedCallback() {
-        window.timingBuffer.push('parent:renderedCallback');
     }
 }
